requirements.txt
fastapi
tortoise-orm[asyncpg]
uvicorn
aerich
asyncpg
python-dotenv
